Updates the portuguese_br translation file.
The file only had 'vessel' translated.

Translation made taking into account game original translation, keeping
where possible translation choices of the game, like:

- Vessel => Hospedeiro
- Dark Bramble => Abrolho Sombrio (translation of names)
- warp => translocation

A special case was made for the "NOMAI_SHUTTLE_COMPUTER" to split the
location with colon and avoid a gender combination ("no", "na") with a
common "em". As locations might have male or female "genders" to combine
with in the Portuguese language.

"Raycast", the name of the mod (New Horizon) and "Dieclone" weren't
translated at all for being either the very name of the mod, or being
considered keywords or neologism words (as much as "Nomai" is not
translated).